Create StatisticsInfo
